requirements.txt
flask==2.3.1
gunicorn==20.1.0
flask-cors==3.0
easyocr==1.7.0
Pillow==9.5.0
requests==2.29.0
google-cloud-storage==2.9.0
google-cloud-secret-manager==2.16.1
